fix(content): address code review safety and type consistency issues
- Use Object.hasOwn() instead of 'in' operator for state config lookup
  to avoid matching inherited properties like toString
- Cast COUNT(*) to int for state_count to maintain consistent types
  with other count expressions in the file

+++ src/lib/server/db/queries/content.ts
@@ -262,7 +262,7 @@ const searchStateSubquery = db
 				SELECT
 					sea.series_id,
 					sr.state,
-					COUNT(*) as state_count,
+					COUNT(*)::int as state_count,
 					ROW_NUMBER() OVER (
 						PARTITION BY sea.series_id
 						ORDER BY CASE sr.state
